Addressed Martins comments
